backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=3.15
# MySQL client - on Windows you may prefer pymysql
# mysqlclient>=2.1.1
pymysql>=1.0.4
